life-os-agent/pyproject.toml
[project]
name = "life-os-agent"
version = "0.1.0"
description = "AI-powered personal time orchestration system"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Kevin Hyde", email = "k.p.hyde@gmail.com" }
]
keywords = ["ai", "scheduling", "calendar", "personal-assistant", "claude"]

dependencies = [
    "anthropic>=0.40.0",
    "google-api-python-client>=2.150.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "pyyaml>=6.0.2",
    "click>=8.1.7",
    "rich>=13.9.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.0",
    "httpx>=0.28.0",
    "mem0ai>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "black>=24.10.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-PyYAML>=6.0.12",
]

[project.scripts]
lifeos = "lifeos.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lifeos"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
